Moving file and sixpack helpers to separate files.
