+++ resources/assets/helpers/file.js
@@ -1,0 +1,54 @@
+/* global Blob, DataView */
+
+import { get } from 'lodash';
+
+/**
+ * Get the type for a specified file.
+ *
+ * @param  {ArrayBuffer} file
+ * @return {String|null}
+ * @todo Eventually deal with other file types.
+ */
+function getFileType(file) {
+  const dv = new DataView(file, 0, 5);
+  const byte1 = dv.getUint8(0, true);
+  const byte2 = dv.getUint8(1, true);
+  const hex = byte1.toString(16) + byte2.toString(16);
+
+  return get(
+    {
+      '8950': 'image/png',
+      '4749': 'image/gif',
+      '424d': 'image/bmp',
+      ffd8: 'image/jpeg',
+    },
+    hex,
+    null,
+  );
+}
+
+/**
+ * Process file (provided as an ArrayBuffer) depending
+ * on its type.
+ *
+ * @param  {ArrayBuffer} file
+ * @return {Blob}
+ *
+ * @todo Eventually deal with other file types.
+ */
+export function processFile(file) {
+  const fileType = getFileType(file);
+  const dataView = new DataView(file);
+
+  if (fileType === 'image/png') {
+    return new Blob([dataView], { type: fileType });
+  }
+
+  if (fileType === 'image/jpeg') {
+    return new Blob([dataView], { type: 'image/jpeg' });
+  }
+
+  throw new Error('Unsupported file type.');
+}
+
+export default null;
