+++ resources/assets/helpers/index.js
@@ -1,8 +1,7 @@
-/* global window, Blob */
+/* global window */
 
 import { format, getTime, isWithinInterval } from 'date-fns';
 import {
-  get,
   isArray,
   isEmpty,
   isNil,
@@ -17,7 +16,6 @@ import {
 
 import { query } from './url';
 import Debug from '../services/Debug';
-import Sixpack from '../services/Sixpack';
 
 /**
  * Return a boolean indicating whether the provided argument is an empty string.
@@ -115,54 +113,6 @@ export function withoutTokens(string) {
  */
 export function pluralize(quantity, singular, plural) {
   return quantity === 1 ? singular : plural;
-}
-
-/**
- * Get the type for a specified file.
- *
- * @param  {ArrayBuffer} file
- * @return {String|null}
- * @todo Eventually deal with other file types.
- */
-function getFileType(file) {
-  const dv = new DataView(file, 0, 5);
-  const byte1 = dv.getUint8(0, true);
-  const byte2 = dv.getUint8(1, true);
-  const hex = byte1.toString(16) + byte2.toString(16);
-
-  return get(
-    {
-      '8950': 'image/png',
-      '4749': 'image/gif',
-      '424d': 'image/bmp',
-      ffd8: 'image/jpeg',
-    },
-    hex,
-    null,
-  );
-}
-
-/**
- * Process file (provided as an ArrayBuffer) depending
- * on its type.
- *
- * @param  {ArrayBuffer} file
- * @return {Blob}
- * @todo Eventually deal with other file types.
- */
-export function processFile(file) {
-  const fileType = getFileType(file);
-  const dataView = new DataView(file);
-
-  if (fileType === 'image/png') {
-    return new Blob([dataView], { type: fileType });
-  }
-
-  if (fileType === 'image/jpeg') {
-    return new Blob([dataView], { type: 'image/jpeg' });
-  }
-
-  throw new Error('Unsupported file type.');
 }
 
 /**
@@ -354,24 +304,6 @@ export function report(error, customAttributes = null) {
   }
 
   window.newrelic.noticeError(errorInstance, customAttributes);
-}
-
-/*
- * Variable that stores single instance of Sixpack.
- */
-let sixpackInstance = null;
-
-/**
- * Get instance of Sixpack class.
- *
- * @return {Sixpack}
- */
-export function sixpack() {
-  if (!sixpackInstance) {
-    sixpackInstance = new Sixpack();
-  }
-
-  return sixpackInstance;
 }
 
 let debugInstance = null;
